add graphql query to generate lazyqueryhook from query
