+++ workspace.json
@@ -9077,6 +9077,13 @@
             "passWithNoTests": true
           },
           "outputs": ["coverage/libs/shared/form-fields"]
+        },
+        "schemas/codegen": {
+          "builder": "@nrwl/workspace:run-commands",
+          "options": {
+            "parallel": false,
+            "command": "graphql-codegen --config libs/shared/form-fields/codegen.yml"
+          }
         }
       }
     },
